+++ inc/roctracer_trace_entries.h
@@ -8,6 +8,7 @@
 #include <roctracer_roctx.h>
 #include <roctracer_hsa.h>
 #include <roctracer_hip.h>
+#include <roctracer_kfd.h>
 #include <ext/prof_protocol.h>
 #include <ext/hsa_rt_utils.hpp>
 
@@ -71,6 +72,18 @@ struct hip_activity_trace_entry_t {
   const roctracer_record_t *record;
   const char *name;
   uint32_t pid;
+};
+
+struct kfd_api_trace_entry_t {
+  std::atomic<uint32_t> valid;
+  roctracer::entry_type_t type;
+  uint32_t domain;
+  uint32_t cid;
+  timestamp_t begin;
+  timestamp_t end;
+  uint32_t pid;
+  uint32_t tid;
+  kfd_api_data_t data;
 };
 
 #endif
